+++ driver/web/apis_admin.go
@@ -236,36 +236,42 @@ func (h AdminAPIsHandler) getSurveys(l *logs.Log, r *http.Request, claims *token
 	}
 	publicStr := r.URL.Query().Get("public")
 
-	var public *bool
+	var public bool
 
 	if publicStr != "" {
-		value, err := strconv.ParseBool(publicStr)
+		valueArchived, err := strconv.ParseBool(publicStr)
 		if err != nil {
 			return l.HTTPResponseErrorAction(logutils.ActionGet, model.TypeSurvey, nil, err, http.StatusInternalServerError, true)
 		}
-		public = &value
+		public = valueArchived
+	} else {
+		valueArchived := false
+		public = valueArchived
 	}
 	archivedStr := r.URL.Query().Get("archived")
 
-	var archived *bool
+	var archived bool
 
 	if archivedStr != "" {
 		valueArchived, err := strconv.ParseBool(archivedStr)
 		if err != nil {
 			return l.HTTPResponseErrorAction(logutils.ActionGet, model.TypeSurvey, nil, err, http.StatusInternalServerError, true)
 		}
-		archived = &valueArchived
+		archived = valueArchived
+	} else {
+		valueArchived := false
+		archived = valueArchived
 	}
 
-	resData, err := h.app.Admin.GetSurveys(claims.OrgID, claims.AppID, nil, surveyIDs, surveyTypes, calendarEventID, &limit, &offset, filter, public, archived)
+	surveys, surverysRsponse, err := h.app.Admin.GetSurveys(claims.OrgID, claims.AppID, &claims.Subject, surveyIDs, surveyTypes, calendarEventID,
+		&limit, &offset, filter, public, archived)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionGet, model.TypeSurvey, nil, err, http.StatusInternalServerError, true)
 	}
 
-	surveys := surveysToSurveyRequests(resData)
-
-	sort.Slice(surveys, func(i, j int) bool {
-		return surveys[i].DateCreated.After(surveys[j].DateCreated)
+	resData := getSurveysResData(surveys, surverysRsponse)
+	sort.Slice(resData, func(i, j int) bool {
+		return resData[i].DateCreated.After(resData[j].DateCreated)
 	})
 
 	rdata, err := json.Marshal(resData)
@@ -282,14 +288,17 @@ func (h AdminAPIsHandler) createSurvey(l *logs.Log, r *http.Request, claims *tok
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionDecode, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
 	}
+	items.CreatorID = claims.Subject
+	items.OrgID = claims.OrgID
+	items.AppID = claims.AppID
+	items.Type = "user"
 
-	item := surveyRequestToSurvey(claims, items)
+	item := surveyRequestToSurvey(items)
 
-	ci, err := h.app.Admin.CreateSurvey(item, claims.ExternalIDs)
+	createdItem, err := h.app.Admin.CreateSurvey(item, claims.ExternalIDs)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionCreate, model.TypeSurvey, nil, err, http.StatusInternalServerError, true)
 	}
-	createdItem := surveyToSurveyRequest(*ci)
 
 	data, err := json.Marshal(createdItem)
 	if err != nil {
@@ -313,7 +322,12 @@ func (h AdminAPIsHandler) updateSurvey(l *logs.Log, r *http.Request, claims *tok
 		return l.HTTPResponseErrorAction(logutils.ActionDecode, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
 	}
 
-	item := updateSurveyRequestToSurvey(claims, items, id)
+	items.CreatorID = claims.Subject
+	items.OrgID = claims.OrgID
+	items.AppID = claims.AppID
+	items.Type = "user"
+
+	item := updateSurveyRequestToSurvey(items, id)
 
 	err = h.app.Admin.UpdateSurvey(item, claims.Subject, claims.ExternalIDs)
 	if err != nil {